Naming changes based on offline discussion.
