+++ packages/android_intent/android/src/main/java/io/flutter/plugins/androidintent/AndroidIntentPluginRegistrar.java
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-package dev.flutter.plugins.androidintent;
+package io.flutter.plugins.androidintent;
 
 import io.flutter.plugin.common.PluginRegistry.Registrar;
