+++ packages/android_intent/android/src/test/java/io/flutter/plugins/androidintent/MethodCallHandlerImplTest.java
@@ -1,4 +1,4 @@
-package dev.flutter.plugins.androidintent;
+package io.flutter.plugins.androidintent;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
