+++ packages/android_intent/example/android/app/src/main/AndroidManifest.xml
@@ -1,14 +1,23 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
   package="io.flutter.plugins.androidintentexample">
 
-  <!-- The INTERNET permission is required for development. Specifically,
-       flutter needs it to communicate with the running application
-       to allow setting breakpoints, to provide hot reload, etc.
-  -->
+  <!-- io.flutter.app.FlutterApplication is an android.app.Application that
+       calls FlutterMain.startInitialization(this); in its onCreate method.
+       In most cases you can leave this as-is, but you if you want to provide
+       additional functionality it is fine to subclass or reimplement
+       FlutterApplication and put your custom class here. -->
   <application
     android:icon="@mipmap/ic_launcher"
     android:label="android_intent_example"
     android:name="io.flutter.app.FlutterApplication">
+    <activity
+      android:configChanges="orientation|keyboardHidden|keyboard|screenSize|locale|layoutDirection"
+      android:exported="true"
+      android:hardwareAccelerated="true"
+      android:launchMode="singleTop"
+      android:name=".EmbeddingV1Activity"
+      android:theme="@android:style/Theme.Black.NoTitleBar"
+      android:windowSoftInputMode="adjustResize"/>
     <activity
       android:configChanges="orientation|keyboardHidden|keyboard|screenSize|locale|layoutDirection"
       android:hardwareAccelerated="true"
@@ -21,21 +30,12 @@
         <category android:name="android.intent.category.LAUNCHER"/>
       </intent-filter>
     </activity>
-    <activity
-      android:configChanges="orientation|keyboardHidden|keyboard|screenSize|locale|layoutDirection"
-      android:exported="true"
-      android:hardwareAccelerated="true"
-      android:launchMode="singleTop"
-      android:name="dev.flutter.plugins.androidintentexample.MainActivity"
-      android:theme="@android:style/Theme.Black.NoTitleBar"
-      android:windowSoftInputMode="adjustResize"/>
   </application>
   <uses-permission android:name="com.android.alarm.permission.SET_ALARM"/>
 
-  <!-- io.flutter.app.FlutterApplication is an android.app.Application that
-       calls FlutterMain.startInitialization(this); in its onCreate method.
-       In most cases you can leave this as-is, but you if you want to provide
-       additional functionality it is fine to subclass or reimplement
-       FlutterApplication and put your custom class here. -->
+  <!-- The INTERNET permission is required for development. Specifically,
+       flutter needs it to communicate with the running application
+       to allow setting breakpoints, to provide hot reload, etc.
+  -->
   <uses-permission android:name="android.permission.INTERNET"/>
 </manifest>
