--- packages/android_intent/example/android/app/src/main/java/dev/flutter/plugins/androidintentexample/MainActivity.java
@@ -1,12 +1,0 @@
-package dev.flutter.plugins.androidintentexample;
-
-import dev.flutter.plugins.androidintent.AndroidIntentPlugin;
-import io.flutter.embedding.android.FlutterActivity;
-import io.flutter.embedding.engine.FlutterEngine;
-
-public class MainActivity extends FlutterActivity {
-  @Override
-  public void configureFlutterEngine(FlutterEngine flutterEngine) {
-    flutterEngine.getPlugins().add(new AndroidIntentPlugin());
-  }
-}
